基础编译通过，定位功能存在bug

+++ app/src/main/java/cn/edu/ecust/myandroidapp/NearbyActivity.kt
@@ -134,7 +134,7 @@ class NearbyActivity : AppCompatActivity() {
                     user.id,
                     currentLat,
                     currentLon,
-                    Constants.MAX_SEARCH_RADIUS.toDouble()
+                    Constants.MAX_SEARCH_RADIUS
                 )
                 
                 allNearbyUsers.clear()
